+++ src/main/resources/assets/minecolonies/lang/manual_en_us.json
@@ -986,6 +986,7 @@
   "com.minecolonies.coremod.gui.workerhuts.recipe.crafting": "Crafting Recipes",
   "com.minecolonies.coremod.gui.workerhuts.recipe.smelting": "Smelting Recipes",
   "com.minecolonies.coremod.gui.workerhuts.recipe.custom": "Custom Recipes",
+  "com.minecolonies.coremod.gui.workerhuts.recipe.domum": "Cutter Recipes",
   "com.minecolonies.coremod.gui.workerhuts.recipe.brewing": "Brewing Recipes",
 
   "item.minecolonies.scan_analyzer": "Schematic Analyzer",
